add product endpoint done

+++ motorcycle-parts-ecommerce/motorPartsHub/src/main/java/edu/icet/repository/ProductImageRepository.java
@@ -3,5 +3,5 @@ package edu.icet.repository;
 import edu.icet.entity.ProductImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
+public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
 }
